Update Main.kt
Refactored the validation for adding and updating cards.
Before validation was an if statement with many statements, now we have an if statement that calls a RarityUtility function that that takes in a string (the rarity entered by the user) and checks if the string is equal to a rarity.

Also we are validating the range of quality. We are calling a Utility function that takes in 3 ints, the input of the user, minimum value and max value. Then the function checks if the input is within the min and max values.