Update Main.kt
Updated the deleteCard function.
This function calls the listAllCards function so the user can see all the cards that they can delete. Then the user is asked to enter the index of the card to delete. A variable called cardToDelete is create and in this variable we call the deleteCard function form the cardAPI. We will use this variable to verify if the card was deleted or not.