updated Main.kt
I have added a load function which basically works the same way as the save function.
The load function calls the load function from CardAPI which takes care of loading the file.
Since we had no card.xml file created, the file was created automatically when we tried saving the files.

The app loads the files on start and saves the files when the user exits the app by entering 0