refactoring and menu
Added the function to the menu, this function is now displayed when the app runs.
The function has been tested and refactored accordingly. The function firstly was displaying cards that "contain" the rarity instead of being "equal" to it. Therefore small refactoring has been implemented.

we will implement testing now.