Update CardAPITest.kt
Testing for the search method has been implemented.
The testing attempts to search for a note that does not have such rarity.
Also tests when a card with a rarity exists. This test tests an existing rarity and a rarity with case that do not match.

All tests have been successful.