Update Main.kt
Implemented the searchByName function.
This function prompts the user to enter a car name which then is passed in to searchByName function in CardAPI which returns the cards that contain the "name" that user have enterd.

Manual testing has been done and the function works perfectly.

Automated testing will be implemented now.