Update Main.kt
Created a new function called listCards() and refactored the menu.

The listCards function is a sub menu from which the user will be able to choose to either list all cards or to list cards by quality.

We will also add the listByRarity function to this menu as it only makes sense for it to be there rather than being a separate option.

No testing yet.