Update Main.kt
New function added. This function displays the number of cards for each rarity.

This function is calls another function form cardAPI which calls the counting functions and that are used in a string.

Manual testing has been done and the app works.
Minor formatting required to make the menu clean.